pyproject.toml
[project]
name = "llm-summarize"
version = "0.1.0"
description = "Ukrainian LLM Summarization with Alignment"
readme = "REPORT.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "bert-score>=0.3.13",
    "datasets==2.21.0",
    "evaluate>=0.4.6",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "hydra-submitit-launcher>=1.2.0",
    "lightning>=2.5.5",
    "peft>=0.17.1",
    "rouge-score>=0.1.2",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "trl[vllm]<0.23.1",
    "vllm>=0.10.2",
    "wandb>=0.22.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "ipython>=9.5.0",
    "jupyter>=1.1.1",
    "pytest>=8.4.2",
    "ruff>=0.13.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
